+++ app/src/main/java/org/linphone/compatibility/ApiTwentyEightPlus.java
@@ -20,19 +20,27 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
 
 import static org.linphone.compatibility.Compatibility.CHAT_NOTIFICATIONS_GROUP;
+import static org.linphone.compatibility.Compatibility.INTENT_LOCAL_IDENTITY;
+import static org.linphone.compatibility.Compatibility.INTENT_MARK_AS_READ_ACTION;
+import static org.linphone.compatibility.Compatibility.INTENT_NOTIF_ID;
+import static org.linphone.compatibility.Compatibility.INTENT_REPLY_NOTIF_ACTION;
+import static org.linphone.compatibility.Compatibility.KEY_TEXT_REPLY;
 
 import android.annotation.TargetApi;
 import android.app.ActivityManager;
 import android.app.Notification;
 import android.app.PendingIntent;
 import android.app.Person;
+import android.app.RemoteInput;
 import android.app.usage.UsageStatsManager;
 import android.content.Context;
+import android.content.Intent;
 import android.graphics.Bitmap;
 import android.graphics.drawable.Icon;
 import org.linphone.R;
 import org.linphone.notifications.Notifiable;
 import org.linphone.notifications.NotifiableMessage;
+import org.linphone.notifications.NotificationBroadcastReceiver;
 
 @TargetApi(28)
 class ApiTwentyEightPlus {
@@ -84,8 +92,8 @@ class ApiTwentyEightPlus {
                 .setShowWhen(true)
                 .setColor(context.getColor(R.color.notification_led_color))
                 .setStyle(style)
-                .addAction(ApiTwentyFourPlus.getReplyMessageAction(context, notif))
-                .addAction(ApiTwentyFourPlus.getMarkMessageAsReadAction(context, notif))
+                .addAction(Compatibility.getReplyMessageAction(context, notif))
+                .addAction(Compatibility.getMarkMessageAsReadAction(context, notif))
                 .build();
     }
 
@@ -113,5 +121,54 @@ class ApiTwentyEightPlus {
                 return "STANDBY_BUCKET_WORKING_SET";
         }
         return null;
+    }
+
+    public static Notification.Action getReplyMessageAction(Context context, Notifiable notif) {
+        String replyLabel = context.getResources().getString(R.string.notification_reply_label);
+        RemoteInput remoteInput =
+                new RemoteInput.Builder(KEY_TEXT_REPLY).setLabel(replyLabel).build();
+
+        Intent replyIntent = new Intent(context, NotificationBroadcastReceiver.class);
+        replyIntent.setAction(INTENT_REPLY_NOTIF_ACTION);
+        replyIntent.putExtra(INTENT_NOTIF_ID, notif.getNotificationId());
+        replyIntent.putExtra(INTENT_LOCAL_IDENTITY, notif.getLocalIdentity());
+
+        PendingIntent replyPendingIntent =
+                PendingIntent.getBroadcast(
+                        context,
+                        notif.getNotificationId(),
+                        replyIntent,
+                        PendingIntent.FLAG_UPDATE_CURRENT);
+
+        return new Notification.Action.Builder(
+                        R.drawable.chat_send_over,
+                        context.getString(R.string.notification_reply_label),
+                        replyPendingIntent)
+                .addRemoteInput(remoteInput)
+                .setAllowGeneratedReplies(true)
+                .setSemanticAction(Notification.Action.SEMANTIC_ACTION_REPLY)
+                .build();
+    }
+
+    public static Notification.Action getMarkMessageAsReadAction(
+            Context context, Notifiable notif) {
+        Intent markAsReadIntent = new Intent(context, NotificationBroadcastReceiver.class);
+        markAsReadIntent.setAction(INTENT_MARK_AS_READ_ACTION);
+        markAsReadIntent.putExtra(INTENT_NOTIF_ID, notif.getNotificationId());
+        markAsReadIntent.putExtra(INTENT_LOCAL_IDENTITY, notif.getLocalIdentity());
+
+        PendingIntent markAsReadPendingIntent =
+                PendingIntent.getBroadcast(
+                        context,
+                        notif.getNotificationId(),
+                        markAsReadIntent,
+                        PendingIntent.FLAG_UPDATE_CURRENT);
+
+        return new Notification.Action.Builder(
+                        R.drawable.chat_send_over,
+                        context.getString(R.string.notification_mark_as_read_label),
+                        markAsReadPendingIntent)
+                .setSemanticAction(Notification.Action.SEMANTIC_ACTION_MARK_AS_READ)
+                .build();
     }
 }