+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Group changes to describe their impact on the project, as follows:
 - Call statistics are now available for each call & conference
 - Added our own devices in LIME encrypted chatrooms' security view
 - No longer display incoming call activity from Service, instead use incoming call notification with full screen intent
+- Improved reply notification when replying to a chat message from the notification
 
 ## [4.1.0] - 2019-05-03